Move precipitation_type directory to precipitation. Many files affected.
