+++ improver/api/__init__.py
@@ -39,7 +39,7 @@ PROCESSING_MODULES = {
     "Combine": "improver.cube_combiner",
     "ConstructReliabilityCalibrationTables": "improver.calibration.reliability_calibration",
     "ContinuousRankedProbabilityScoreMinimisers": "improver.calibration.ensemble_calibration",
-    "ConvectionRatioFromComponents": "improver.precipitation_type.convection",
+    "ConvectionRatioFromComponents": "improver.precipitation.convection",
     "ConvertProbabilitiesToPercentiles": "improver.ensemble_copula_coupling.ensemble_copula_coupling",
     "CopyMetadata": "improver.utilities.copy_metadata",
     "CorrectLandSeaMask": "improver.generate_ancillaries.generate_ancillary",
@@ -58,7 +58,7 @@ PROCESSING_MODULES = {
     "ExtractValueFromTable": "improver.utilities.extract_from_table",
     "FieldTexture": "improver.utilities.textural",
     "FillRadarHoles": "improver.nowcasting.utilities",
-    "FreezingRain": "improver.precipitation_type.freezing_rain",
+    "FreezingRain": "improver.precipitation.freezing_rain",
     "FrictionVelocity": "improver.wind_calculations.wind_downscaling",
     "GenerateClearskySolarRadiation": "improver.generate_ancillaries.generate_derived_solar_fields",
     "GenerateOrographyBandAncils": "improver.generate_ancillaries.generate_ancillary",
@@ -67,7 +67,7 @@ PROCESSING_MODULES = {
     "GenerateTopographicZoneWeights": "improver.generate_ancillaries.generate_topographic_zone_weights",
     "GradientBetweenAdjacentGridSquares": "improver.utilities.spatial",
     "GradientBetweenVerticalLevels": "improver.utilities.gradient_between_vertical_levels",
-    "HailFraction": "improver.precipitation_type.hail_fraction",
+    "HailFraction": "improver.precipitation.hail_fraction",
     "HailSize": "improver.psychrometric_calculations.hail_size",
     "height_of_maximum": "improver.utilities.cube_manipulation",
     "HumidityMixingRatio": "improver.psychrometric_calculations.psychrometric_calculations",
@@ -106,10 +106,10 @@ PROCESSING_MODULES = {
     "ResolveWindComponents": "improver.wind_calculations.wind_components",
     "RoughnessCorrection": "improver.wind_calculations.wind_downscaling",
     "SaturatedVapourPressureTable": "improver.generate_ancillaries.generate_svp_table",
-    "ShowerConditionProbability": "improver.precipitation_type.shower_condition_probability",
+    "ShowerConditionProbability": "improver.precipitation.shower_condition_probability",
     "SignificantPhaseMask": "improver.psychrometric_calculations.significant_phase_mask",
-    "SnowFraction": "improver.precipitation_type.snow_fraction",
-    "SnowSplitter": "improver.precipitation_type.snow_splitter",
+    "SnowFraction": "improver.precipitation.snow_fraction",
+    "SnowSplitter": "improver.precipitation.snow_splitter",
     "SpatiallyVaryingWeightsFromMask": "improver.blending.spatial_weights",
     "SpotExtraction": "improver.spotdata.spot_extraction",
     "SpotHeightAdjustment": "improver.spotdata.height_adjustment",
