+++ improver/cli/freezing_rain.py
@@ -39,6 +39,6 @@ def process(*cubes: cli.inputcube, model_id_attr: str = None):
             A cube of freezing rain rate or accumulation probabilities.
 
     """
-    from improver.precipitation_type.freezing_rain import FreezingRain
+    from improver.precipitation.freezing_rain import FreezingRain
 
     return FreezingRain(model_id_attr=model_id_attr)(*cubes)
