+++ improver/precipitation/hail_fraction.py
