+++ improver_tests/precipitation/calculate_sleet_prob/test_calculate_sleet_probability.py
@@ -9,7 +9,7 @@ import unittest
 import numpy as np
 from iris.tests import IrisTest
 
-from improver.precipitation_type.calculate_sleet_prob import calculate_sleet_probability
+from improver.precipitation.calculate_sleet_prob import calculate_sleet_probability
 from improver.synthetic_data.set_up_test_cubes import set_up_probability_cube
