+++ improver_tests/precipitation/shower_condition_probability/test_ShowerConditionProbability.py
@@ -12,7 +12,7 @@ from iris.cube import CubeList
 from numpy import ndarray
 
 from improver.metadata.constants import FLOAT_DTYPE
-from improver.precipitation_type.shower_condition_probability import (
+from improver.precipitation.shower_condition_probability import (
     ShowerConditionProbability,
 )
 from improver.synthetic_data.set_up_test_cubes import set_up_variable_cube
